git-svn-id: https://openil.svn.sourceforge.net/svnroot/openil/branches/DevIL2@1643 e624367d-d6e4-4977-a939-15c673ed3205

+++ src-ILU/src/ilu_rotate.cpp
@@ -120,7 +120,7 @@ ILAPI ILimage* ILAPIENTRY iluRotate_(ILimage *Image, ILfloat Angle)
 		return NULL;
 	}
 
-	if (ilResizeImage(Rotated, (ILuint)ceil(fabs(MaxX) - MinX), (ILuint)ceil(fabs(MaxY) - MinY), 1, Image->Bpp, Image->Bpc) == IL_FALSE) {
+	if (ilResizeImage(Rotated, (ILuint)ceil(fabs((float)MaxX) - MinX), (ILuint)ceil(fabs((float)MaxY) - MinY), 1, Image->Bpp, Image->Bpc) == IL_FALSE) {
 		ilCloseImage(Rotated);
 		return IL_FALSE;
 	}
